fix: convert BigInt types (i64/u64/usize) to i32/u32/f64 for specta-typescript 0.0.12 compatibility; update env_filter 0.1.4→2.0.0, bump JS deps (lucide-react 1.x, react-i18next 17.x, zod 4.x, etc.)

+++ src-tauri/src/commands/history.rs
@@ -11,8 +11,8 @@ use tauri::{AppHandle, State};
 pub async fn get_history_entries(
     _app: AppHandle,
     history_manager: State<'_, Arc<HistoryManager>>,
-    cursor: Option<i64>,
-    limit: Option<usize>,
+    cursor: Option<i32>,
+    limit: Option<u32>,
 ) -> Result<PaginatedHistory, String> {
     history_manager
         .get_history_entries(cursor, limit)
@@ -25,7 +25,7 @@ pub async fn get_history_entries(
 pub async fn toggle_history_entry_saved(
     _app: AppHandle,
     history_manager: State<'_, Arc<HistoryManager>>,
-    id: i64,
+    id: i32,
 ) -> Result<(), String> {
     history_manager
         .toggle_saved_status(id)
@@ -51,7 +51,7 @@ pub async fn get_audio_file_path(
 pub async fn delete_history_entry(
     _app: AppHandle,
     history_manager: State<'_, Arc<HistoryManager>>,
-    id: i64,
+    id: i32,
 ) -> Result<(), String> {
     history_manager
         .delete_entry(id)
@@ -65,7 +65,7 @@ pub async fn retry_history_entry_transcription(
     app: AppHandle,
     history_manager: State<'_, Arc<HistoryManager>>,
     transcription_manager: State<'_, Arc<TranscriptionManager>>,
-    id: i64,
+    id: i32,
 ) -> Result<(), String> {
     let entry = history_manager
         .get_entry_by_id(id)
@@ -111,7 +111,7 @@ pub async fn retry_history_entry_transcription(
 pub async fn update_history_limit(
     app: AppHandle,
     history_manager: State<'_, Arc<HistoryManager>>,
-    limit: usize,
+    limit: u32,
 ) -> Result<(), String> {
     let mut settings = crate::settings::get_settings(&app);
     settings.history_limit = limit;
